requirements.txt
azure-common==1.1.27
azure-core==1.13.0
azure-functions==1.7.0
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
isodate==0.6.0
msal==1.10.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.0
portalocker==1.7.1
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.17.0
#pywin32==300
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
